+++ mitra_cli/src/commands/mod.rs
@@ -2,5 +2,6 @@ pub mod account;
 pub mod activitypub;
 pub mod config;
 pub mod filter;
+pub mod invoice;
 pub mod post;
 pub mod storage;